requirements.txt
PySide6
Pillow 